extend doc-examples test to client-directive.md + fragment.md (#1439)
v1 covered a single doc page (`jsx-compatibility.md`). This expands the
mechanism across the rest of `docs/core/rendering/` by parameterising
the test over a list of `PageSpec` objects sharing one scaffold.

Extractor changes:
- Also split on blank lines, so adjacent unmarked snippets in a single
  fence become separate test cases instead of being combined into one
  invalid body. v1 behaviour on `jsx-compatibility.md` is unchanged
  (15 pass / 6 skip, exact match).
- `PageSpec.pageSkip` per-page predicate for opting out illustrative
  snippets (e.g. `{/* @client */ expression}` placeholder).

Scaffold:
- Added `children` via `props.children` so fragment.md's
  `<>{children}</>` examples compile inside the wrapper `<div>`.

Coverage:
- jsx-compatibility.md: 21 tests, 15 pass / 6 skip (unchanged)
- client-directive.md: 6 tests, 4 pass / 2 skip (new)
- fragment.md: 3 tests, 3 pass / 0 skip (new)
- Total: 30 tests, 22 pass / 8 skip / 0 fail

File renamed from `doc-examples-jsx-compatibility.test.ts` to
`doc-examples.test.ts` to reflect the multi-page scope.
